Initial commit
Branched off from PatternLab package.

This package contains the resources and lexicons of the pattern library, which might be accompanied later by a few small components that require to be installed through a package (such as custom CB inputs).

In contrary to Romanesco Patterns, this package will be a required dependency in all Romanesco Soil installs.