Import assets folder from Soil repo
Now that the Romanesco Soil repository is supposed to be used as a boilerplate only, the assets need to be included somewhere else, so the can still receive updates. Backyard seems to be the best location.

The assets are not referenced correctly in their new location yet at this point.